@@ -1,9 +1,13 @@
 <script lang="ts">
 	import { signInWith } from '$lib/modules/firebase';
 
-	export let ping: boolean;
-	export let logged: boolean;
-	export let picture: string | undefined;
+	interface Props {
+		ping: boolean;
+		logged: boolean;
+		picture?: string;
+	}
+
+	let { ping, logged, picture }: Props = $props();
 </script>
 
 {#if logged}
@@ -13,13 +17,13 @@
 			<span class="absolute top-0 right-0 flex h-3 w-3">
 				<span
 					class="animate-ping absolute inline-flex h-full w-full rounded-full bg-sky-400 opacity-75"
-				/>
-				<span class="relative inline-flex rounded-full h-3 w-3 bg-sky-500" />
+				></span>
+				<span class="relative inline-flex rounded-full h-3 w-3 bg-sky-500"></span>
 			</span>
 		{/if}
 	</a>
 {:else}
-	<button on:click={() => signInWith('google')} class="h-full">
+	<button onclick={() => signInWith('google')} class="h-full">
 		<img
 			src="https://lh3.googleusercontent.com/COxitqgJr1sJnIDe8-jiKhxDx1FrYbtRHKJ9z_hELisAlapwE9LUPh6fcXIfb5vwpbMl4xl9H9TRFPc5NOO8Sb3VSgIBrfRYvW6cUA"
 			alt="Login"
