@@ -1,8 +1,13 @@
 <script lang="ts">
 	import '../app.css';
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
 </script>
 
-<slot />
+{@render children?.()}
 
 <style lang="postcss">
 	@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
